# Improve memory usage of enrichTags (dogstatsd)