--- a/pkg/dogstatsd/enrich.go
+++ b/pkg/dogstatsd/enrich.go
@@ -25,6 +25,7 @@ var (
 )
 
 func enrichTags(tags []string, defaultHostname string, originTagsFunc func() []string, entityIDPrecedenceEnabled bool) ([]string, string) {
+	var originTags, entityTags []string
 	host := defaultHostname
 
 	n := 0
@@ -40,9 +41,10 @@ func enrichTags(tags []string, defaultHostname string, originTagsFunc func() []s
 		}
 	}
 	tags = tags[:n]
+
 	if entityIDValue == "" || !entityIDPrecedenceEnabled {
 		// Add origin tags only if the entity id tags is not provided
-		tags = append(tags, originTagsFunc()...)
+		originTags = originTagsFunc()
 	}
 	if entityIDValue != "" && entityIDValue != entityIDIgnoreValue {
 		// Check if the value is not "none" in order to avoid calling
@@ -50,16 +52,29 @@ func enrichTags(tags []string, defaultHostname string, originTagsFunc func() []s
 
 		// currently only supported for pods
 		entity := kubelet.KubePodTaggerEntityPrefix + entityIDValue
-		entityTags, err := getTags(entity, tagger.DogstatsdCardinality)
+		var err error
+		entityTags, err = getTags(entity, tagger.DogstatsdCardinality)
 		if err != nil {
 			log.Tracef("Cannot get tags for entity %s: %s", entity, err)
 			tlmUDPOriginDetectionError.Inc()
-		} else {
-			tags = append(tags, entityTags...)
 		}
 	}
-	return tags, host
+
+	var finalTags []string
+	if cap(tags) >= len(tags)+len(originTags)+len(entityTags) {
+		tags = append(tags, originTags...)
+		tags = append(tags, entityTags...)
+		finalTags = tags
+	} else {
+		finalTags = make([]string, 0, len(tags)+len(originTags)+len(entityTags))
+		finalTags = append(finalTags, tags...)
+		finalTags = append(finalTags, originTags...)
+		finalTags = append(finalTags, entityTags...)
+	}
+
+	return finalTags, host
 }
+
 func enrichMetricType(dogstatsdMetricType metricType) metrics.MetricType {
 	switch dogstatsdMetricType {
 	case gaugeType: