### What does this PR do?

Small reduction in dogstatsd memory usage in `enrichTags`:

Before:
```
BenchmarkEnrichTags
BenchmarkEnrichTags/20-tags
BenchmarkEnrichTags/20-tags-8            7303458           172 ns/op         224 B/op          1 allocs/op
BenchmarkEnrichTags/40-tags
BenchmarkEnrichTags/40-tags-8            4497426           261 ns/op         416 B/op          1 allocs/op
BenchmarkEnrichTags/60-tags
BenchmarkEnrichTags/60-tags-8            3551148           334 ns/op         640 B/op          1 allocs/op
BenchmarkEnrichTags/80-tags
BenchmarkEnrichTags/80-tags-8            2889858           414 ns/op         896 B/op          1 allocs/op
BenchmarkEnrichTags/100-tags
BenchmarkEnrichTags/100-tags-8           2305017           507 ns/op        1024 B/op          1 allocs/op
BenchmarkEnrichTags/120-tags
BenchmarkEnrichTags/120-tags-8           2100564           580 ns/op        1280 B/op          1 allocs/op
BenchmarkEnrichTags/140-tags
BenchmarkEnrichTags/140-tags-8           1845979           625 ns/op        1408 B/op          1 allocs/op
BenchmarkEnrichTags/160-tags
BenchmarkEnrichTags/160-tags-8           1474875           822 ns/op        1792 B/op          1 allocs/op
BenchmarkEnrichTags/180-tags
BenchmarkEnrichTags/180-tags-8           1470181           802 ns/op        1792 B/op          1 allocs/op
BenchmarkEnrichTags/200-tags
BenchmarkEnrichTags/200-tags-8           1367126           903 ns/op        2048 B/op          1 allocs/op
```

After:
```
BenchmarkEnrichTags/20-tags
BenchmarkEnrichTags/20-tags-8            7106082           176 ns/op         224 B/op          1 allocs/op
BenchmarkEnrichTags/40-tags
BenchmarkEnrichTags/40-tags-8            4521068           255 ns/op         416 B/op          1 allocs/op
BenchmarkEnrichTags/60-tags
BenchmarkEnrichTags/60-tags-8            3717309           331 ns/op         640 B/op          1 allocs/op
BenchmarkEnrichTags/80-tags
BenchmarkEnrichTags/80-tags-8            2713755           420 ns/op         896 B/op          1 allocs/op
BenchmarkEnrichTags/100-tags
BenchmarkEnrichTags/100-tags-8           2346262           517 ns/op        1024 B/op          1 allocs/op
BenchmarkEnrichTags/120-tags
BenchmarkEnrichTags/120-tags-8           2094024           610 ns/op        1280 B/op          1 allocs/op
BenchmarkEnrichTags/140-tags
BenchmarkEnrichTags/140-tags-8           1826889           631 ns/op        1408 B/op          1 allocs/op
BenchmarkEnrichTags/160-tags
BenchmarkEnrichTags/160-tags-8           1490360           811 ns/op        1792 B/op          1 allocs/op
BenchmarkEnrichTags/180-tags
BenchmarkEnrichTags/180-tags-8           1589427           757 ns/op        1792 B/op          1 allocs/op
BenchmarkEnrichTags/200-tags
BenchmarkEnrichTags/200-tags-8           1419412           882 ns/op        2048 B/op          1 allocs/op
```

### Motivation

What inspired you to submit this pull request?

### Additional Notes

Anything else we should know when reviewing?

### Describe your test plan

Performance optimisation, nothing should change before/after.