@vboulineau 2020-09-18
> Can you update the benchmark numbers in the PR description?
> 
> I'd also be interested in how the change affects the speed of the function: since the number of allocation hasn't changed, I wonder if the total size allocated matters that much, and if it's visible in the raw speed of the function. Maybe the code changes actually make the function slower for some reason.
> 
> In any case, if the updated benchmarks look good, PR LGTM 👍

Updated bench to be in a situation where old code actually needs to re-allocate, updated results in PR, still no diff in the bench though I see with the debugger that `tags` change size twice. Maybe there's some smart reuse of previously allocated blocs, I don't know.
Worst case change does not hurt, best case it brings a little.